+++ R/accounts.R
@@ -486,7 +486,9 @@ accountInfoFromHostUrl <- function(hostUrl) {
 
 
 secret <- function(x) {
-  stopifnot(is.character(x))
+  if (is.null(x)) return(NULL)
+
+  stopifnot(is.character(x) || all(is.na(x)))
   structure(x, class = "rsconnect_secret")
 }